P0088R3_variant_msvc: Extract, part 1. Pure movement, no changes.
